adjust logic for failed start hooks

+++ src/events/etcd.py
@@ -132,8 +132,8 @@ class EtcdEvents(Object):
         ):
             # this unit has been added to the etcd cluster
             if not self.charm.state.cluster.auth_enabled:
+                # failed start hooks on cluster initialization will go here on retry
                 if self.charm.unit.is_leader():
-                    # if enabling auth failed on first cluster startup, we want to retry now
                     try:
                         self.charm.cluster_manager.enable_authentication()
                         self.charm.state.cluster.update({"authentication": "enabled"})
@@ -143,19 +143,21 @@ class EtcdEvents(Object):
                 else:
                     raise EtcdAuthNotEnabledError("Authentication not enabled.")
 
-            if self.charm.workload.exists(DATABASE_DIR):
-                logger.warning(f"Existing database file detected in {DATABASE_DIR}.")
-                # storage cannot be reused on non-leader members
-                try:
-                    self.charm.workload.remove_directory(DATABASE_DIR)
-                    logger.warning(
-                        f"Removed database file from {DATABASE_DIR} to join existing cluster."
-                    )
-                except OSError:
-                    # if removing fails, we cannot start the workload or the member would crash
-                    raise
+            if not self.charm.state.unit_server.is_started:
+                # database files should not be deleted on running units
+                if self.charm.workload.exists(DATABASE_DIR):
+                    logger.warning(f"Existing database file detected in {DATABASE_DIR}.")
+                    # storage cannot be reused on non-leader members
+                    try:
+                        self.charm.workload.remove_directory(DATABASE_DIR)
+                        logger.warning(
+                            f"Removed database file from {DATABASE_DIR} to join existing cluster."
+                        )
+                    except OSError:
+                        # if removing fails, we cannot start the workload or the member would crash
+                        raise
 
-            self.charm.cluster_manager.start_member()
+                self.charm.cluster_manager.start_member()
         else:
             # this unit that has not yet been added to the cluster
             # wait for leader to process `relation_joined` event and add the member to the cluster
